pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hyperliquid-trading-bot"
version = "0.1.0"
description = "AI-powered async trading bot for Hyperliquid DEX using LangChain"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "nubro999", email = "zcb167@gmail.com"}
]
keywords = ["trading", "bot", "cryptocurrency", "hyperliquid", "langchain", "async"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core async
    "aiohttp>=3.9.0",
    "websockets>=12.0",
    "asyncio>=3.4.3",

    # LangChain stack
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langchain-community>=0.0.20",
    "langgraph>=0.0.20",
    "langsmith>=0.0.80",

    # Blockchain
    "eth-account>=0.11.0",
    "web3>=6.15.0",

    # Data & Storage
    "chromadb>=0.4.22",
    "redis>=5.0.1",
    "influxdb-client>=1.40.0",

    # Data processing
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "pydantic>=2.6.0",

    # Configuration
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",

    # Monitoring & Logging
    "loguru>=0.7.2",
    "rich>=13.7.0",
    "prometheus-client>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]

[project.urls]
Homepage = "https://github.com/nubro999/async_rag_py"
Documentation = "https://github.com/nubro999/async_rag_py"
Repository = "https://github.com/nubro999/async_rag_py"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "B", "C90"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=html --cov-report=term"
